+++ javascript/node/selenium-webdriver/lib/command.js
@@ -232,26 +232,6 @@ class Executor {
 }
 
 
-/**
- * Wraps a promised {@link Executor}, ensuring no commands are executed until
- * the wrapped executor has been fully resolved.
- * @implements {Executor}
- */
-class DeferredExecutor {
-  /**
-   * @param {!Promise<Executor>} delegate The promised delegate, which
-   *     may be provided by any promise-like thenable object.
-   */
-  constructor(delegate) {
-
-    /** @override */
-    this.execute = function(command) {
-      return delegate.then(executor => executor.execute(command));
-    };
-  }
-}
-
-
 
 // PUBLIC API
 
@@ -259,6 +239,5 @@ class DeferredExecutor {
 module.exports = {
   Command: Command,
   Name: Name,
-  Executor: Executor,
-  DeferredExecutor: DeferredExecutor
+  Executor: Executor
 };